// src/data/model/packet/PacketDataType.ts
export enum PACKET_DATA_TYPES {

    GENERAL_CONNECTION_SUCCEED = "GENERAL_CONNECTION_SUCCEED",

    AUTH_SING_IN = "AUTH_SING_IN",
    AUTH_SING_UP = "AUTH_SING_UP",
    AUTH_SING_OUT = "AUTH_SING_OUT",

    PLACE_CREATE = "PLACE_CREATE",
    PLACE_GET = "PLACE_GET",
    PLACE_GET_MY = "PLACE_GET_MY",
    PLACE_DELETE = "PLACE_DELETE",
    PLACE_UPDATE = "PLACE_UPDATE",
    PLACE_JOIN = "PLACE_JOIN",
    PLACE_LEAVE = "PLACE_LEAVE",

    ROLE_CREATE = "ROLE_CREATE",
    ROLE_GET = "ROLE_GET",
    ROLE_DELETE = "ROLE_DELETE",
    ROLE_UPDATE = "ROLE_UPDATE",

    LAW_CREATE = "LAW_CREATE",
    LAW_GET = "LAW_GET",
    LAW_DELETE = "LAW_DELETE",
    LAW_UPDATE = "LAW_UPDATE",

    PROFILE_UPDATE_MUTABLE_DATA = "PROFILE_UPDATE_MUTABLE_DATA",

}